Factor out Tile as a separate entity
