+++ src/test/java/org/imintel/mbtiles4j/MBTileReaderTest.java
@@ -1,13 +1,22 @@
 package org.imintel.mbtiles4j;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
 import org.imintel.mbtiles4j.model.MetadataBounds;
 import org.imintel.mbtiles4j.model.MetadataEntry;
 import org.junit.Test;
-
-import java.io.File;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 public class MBTileReaderTest {
